[CoreEngine] fixed the issue that the gpu id is released with no properly cases.
